requirements.txt
pandas
numpy
dvc
keras
scikit-learn
seaborn
tensorflow
joblib
PyYaml
streamlit
keras_preprocessing
mlflow
h5py
opencv-python
